add csidriver_deployed.go, refactor etcdcluster_created.go
