+++ pkg/controller/glustercluster/etcdcluster_created.go
@@ -1,0 +1,27 @@
+package glustercluster
+
+import (
+	"github.com/gluster/anthill/pkg/reconciler"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+)
+
+var etcdClusterCreated = reconciler.NewAction(
+	"etcdClusterCreated",
+	[]*reconciler.Action{
+		etcdCRDExists,
+	},
+	func(request reconcile.Request, client client.Client, scheme *runtime.Scheme) (reconciler.Result, error) {
+		return reconciler.Result{Status: corev1.ConditionTrue, Message: "it's true"}, nil
+	},
+)
+
+var etcdCRDExists = reconciler.NewAction(
+	"etcdCRDExists",
+	[]*reconciler.Action{},
+	func(request reconcile.Request, client client.Client, scheme *runtime.Scheme) (reconciler.Result, error) {
+		return reconciler.Result{Status: corev1.ConditionTrue, Message: "it's true"}, nil
+	},
+)
